travelappbackend/requirements.txt
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
psycopg2-binary==2.9.3
Flask-JWT-Extended==4.4.3
Werkzeug==2.0.3
PyJWT==2.3.0
gunicorn==20.1.0
